+++ src/main/java/io/github/alexiscomete/lapinousecond/useful/transactions/Offer.kt
@@ -13,8 +13,8 @@ val offers = CacheCustom(OFFERS) { Offer(it) }
 class Offer(id: Long) : CacheGetSet(id, OFFERS), Transaction {
     override val who: Player
         get() = players[this["who"].toLong()]!!
-    override val amount: Int
-        get() = this["amount"].toInt()
+    override val amount: Double
+        get() = this["amount"].toDouble()
     override val what: Resource
         get() = Resource.valueOf(this["what"].uppercase())
     override val amountRB: Double
