+++ src/main/java/io/github/alexiscomete/lapinousecond/useful/transactions/Research.kt
@@ -13,8 +13,8 @@ val researches = CacheCustom(RESEARCHES) { Research(it) }
 class Research(id: Long) : CacheGetSet(id, RESEARCHES), Transaction {
     override val who: Player
         get() = players[this["who"].toLong()]!!
-    override val amount: Int
-        get() = this["amount"].toInt()
+    override val amount: Double
+        get() = this["amount"].toDouble()
     override val what: Resource
         get() = Resource.valueOf(this["what"].uppercase())
     override val amountRB: Double
